tests: every fake release download now hands out a signature (Remote download, SSH deploy e2e)
The precompiled cache verifies the archive's .minisig on download and on
every hit, so the two other tests that fake _download — the FrameOS Remote
download and the real-SSH deploy e2e's locally packaged release — sign their
archives with the shared test key (release_signing_helpers.py) and trust it.
